Renaming ext/AE/AE.c to ext/ae/ae.c + necessary changes to support
Mostly following the casing / naming conventions as described in:

  https://guides.rubygems.org/gems-with-extensions/

Namely, files are lowercase, the AE module is top-level (eg not
AE/AE), and lib/ae.rb is removed and we load the bundle directly.

This fixes a number of build issues I've been having.

Directly tested by installing the gem privately and then ran code that
must load the bundle.